\docType{data}
\name{alexfiord}
\alias{alexfiord}
\title{Alexandra Fiord transect dataset}
\format{A data frame with 93 rows and 10 variables.}
\source{
 Stewart, K. J., E. G. Lamb, D. S. Coxon, and S. D. Siciliano. 2011b. Bryophyte-cyanobacterial associations as a key factor in N2-fixation across the Canadian Arctic. \emph{Plant and Soil} \bold{ 344}:335-346.
}
\description{
  This dataset summarizes ground cover and nitrogen fixation activity along
  a 93 point variably spaced transect at Alexandra Fiord, Ellesmere Island, 
  Nunavut, Canada.
}
\details{
  Variables are: X, Y, Moisture, N_Fix, SoilCrust, Bryoph, 
  Lich, Forbs, Gram, Shrubs
}
\keyword{datasets}