\docType{data}
\name{plantcomp}
\alias{plantcomp}
\title{Plant Competition dataset}
\format{A data frame with 192 rows and 10 variables.}
\source{
Lamb, E. G. and J. F. Cahill. 2008. When competition does not matter: grassland diversity and community composition. \emph{American Naturalist} \bold{171}:777-787
}
\description{
  This dataset summarizes plant community characteristics, competition
  intensity, and topographic features in a mesic grassland in central
  Alberta, Canada.
}
\details{
 Variables are: Easting ,Northing ,Comp.Intensity ,ShootBio ,RootBio ,SpRich ,
  SoilMoist ,LightInt ,Topog.Pos ,TotalN 
}
\keyword{datasets}