\docType{data}
\name{truelove}
\alias{truelove}
\title{Truelove lowland transect dataset}
\format{A data frame with 129 rows and 10 variables.}
\source{
 Stewart, K. J., E. G. Lamb, D. S. Coxon, and S. D. Siciliano. 2011. Bryophyte-cyanobacterial associations as a key factor in N2-fixation across the Canadian Arctic. \emph{Plant and Soil} \bold{ 344}:335-346.
}
\description{
  This dataset summarizes ground cover and nitrogen fixation activity along
  a 129 point regularly spaced transect on Truelove Lowland, Devon Island, Nunavut, Canada.
}
\details{
  Variables are: X, Y, Moisture, N_Fix, SoilCrust, Bryoph, 
  Lich, Forbs, Gram, Shrubs
}
\keyword{datasets}